requirements.txt
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.4
numpy==1.26.2

# NLP and ML
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2
Pillow==10.1.0

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Scraping
selenium==4.15.2
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

